Move LocaleProxy to its own file like the others.
